% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{as_viewr}
\alias{as_viewr}
\title{Convert data from another format into a viewr object}
\usage{
as_viewr(
  obj_name,
  frame_rate = 100,
  frame,
  time_sec,
  subject,
  position_length,
  position_width,
  position_height,
  include_rotation = FALSE,
  rotation_real,
  rotation_length,
  rotation_width,
  rotation_height
)
}
\arguments{
\item{obj_name}{A tibble or data frame containing movement trajectories}

\item{frame_rate}{Must be a single numeric value indicating capture frame
rate in frames per second.}

\item{frame}{Column number of obj_name that contains frame numbers}

\item{time_sec}{Column number of obj_name that contains time (must be in
seconds)}

\item{subject}{Column number of obj_name that contains subject name(s)}

\item{position_length}{Column number of obj_name that contains length-axis
position values}

\item{position_width}{Column number of obj_name that contains width-axis
position values}

\item{position_height}{Column number of obj_name that contains height-axis
postion values}

\item{include_rotation}{Are rotation data included? Defaults to FALSE}

\item{rotation_real}{Column number of obj_name that contains the "real" axis
of quaternion rotation data}

\item{rotation_length}{Column number of obj_name that contains the length
axis of quaternion rotation data}

\item{rotation_width}{Column number of obj_name that contains the width axis
of quaternion rotation data}

\item{rotation_height}{Column number of obj_name that contains the height
axis of quaternion rotation data}
}
\value{
A tibble that is organized to be compliant with other
\code{pathviewR} functions and that contains the attributes
\code{pathviewR_steps} with entries set to \code{c("viewr",
  "renamed_tunnel", "gathered_tunnel")}
}
\description{
Convert data from another format into a viewr object
}
\examples{

## Create a dummy data frame
df <- data.frame(frame = seq(1, 100, by = 1),
                 time_sec = seq(0, by = 0.01, length.out = 100),
                 subject = "bob",
                 z = rnorm(100),
                 x = rnorm(100),
                 y = rnorm(100))

## Use as_viewr() to convert it into a viewr object
test <-
  as_viewr(
    df,
    frame = 1,
    time_sec = 2,
    subject = 3,
    position_length = 5,
    position_width = 6,
    position_height = 4
  )
}
\author{
Vikram B. Baliga
}
