% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{bind_viewr_objects}
\alias{bind_viewr_objects}
\title{Bind viewr objects}
\usage{
bind_viewr_objects(obj_list)
}
\arguments{
\item{obj_list}{A list of viewr objects}
}
\value{
A single viewr object (tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}) that combines all the
rows of the source viewr objects in \code{obj_list}. Metadata may not
necessarily be retained and therefore \code{attributes} should be used with
caution.
}
\description{
Combine a list of multiple viewr objects into a single viewr object
}
\seealso{
Other batch functions: 
\code{\link{clean_viewr_batch}()},
\code{\link{import_and_clean_batch}()},
\code{\link{import_batch}()}
}
\author{
Vikram B. Baliga
}
\concept{batch functions}
