% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{calc_vis_angle_box}
\alias{calc_vis_angle_box}
\title{Estimate visual angles in a box-shaped tunnel}
\usage{
calc_vis_angle_box(obj_name, simplify_output = FALSE)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}

\item{simplify_output}{If TRUE, the returned object includes minimum distance
to lateral walls as well as the visual angles (in degrees) to those points on
the walls. If FALSE, the returned object includes all variables internal
to the calculation.}
}
\value{
A tibble or data.frame with added variables for
\code{vis_angle_pos_deg} and \code{vis_angle_neg_deg} reported in degrees.
}
\description{
Based on rigid body, i.e animal head positions in a b-shaped tunnel,
\code{calc_vis_angle_box()} calculates the visual angles created by lateral
visual stimuli.
}
\details{
\code{cal_vis_angle_box} assumes fixed gaze at the point on the
either side of the tunnel that minimizes the distance to visual stimuli and
thereby maximizes visual angles.
All legth measurements reported in meters.
}
\examples{

}
\seealso{
Other visual perception functions: 
\code{\link{calc_vis_angle_V}()}
}
\author{
Eric R. Press
}
\concept{visual perception functions}
