% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{calc_vis_angle_mod}
\alias{calc_vis_angle_mod}
\title{Based on rigid body, i.e animal head positions in a "V" shaped tunnel,
\code{calc_vis_angle_mod()} calculates the visual angles created by lateral
visual stimuli.}
\usage{
calc_vis_angle_mod(
  obj_name,
  vertex_angle = 45,
  gnd_plane,
  stim_param_pos,
  stim_param_neg
)
}
\arguments{
\item{obj_name}{A tibble or data.frame with attribute \code{viewr}}

\item{vertex_angle}{The angle (in degrees) subtended by a vertical axis and
the sides of the tunnel. Equivalent to the angle of the "V" divided by 2.
\code{vertex_angle)} defaults to 45.}

\item{gnd_plane}{The vertical distance (in meters) between the bottom of the
"V" and the horizontal plane through the origin (0,0,0).}

\item{stim_param_pos}{The length (in meters) of the visual stimulus presented
on the positive side of the tunnel (i.e. \code{Position_widths >= 0}). For
example, a sinusoidal grating 10cm wide is \code{stim_param_pos = 0.1}}

\item{stim_param_neg}{The same convention as \code{stim_param_pos} but for
stimuli presented on the negative side of the tunnel
(i.e. \code{Position_widths < 0}).}
}
\value{
A tibble or data.frame with added variables for
\code{vis_angle_pos_deg} and \code{vis_angle_neg_deg} reported in degrees.
}
\description{
Based on rigid body, i.e animal head positions in a "V" shaped tunnel,
\code{calc_vis_angle_mod()} calculates the visual angles created by lateral
visual stimuli.
}
\details{
\code{cal_vis_angle_mod} assumes fixed gaze at the point on the
either side of the tunnel that minimizes the distance to visual stimuli and
thereby maximizes visual angles. Currently, visual angles are overestimated
when \code{Position_widths} lies outside the boundaries starting at the
bottom of the "V" and extend at right angles from the planes created by
either sideof the tunnel. As \code{vertex_angle} increases from 45 t0 90,
this boundary where calc_vis_angle_mod is accurate becomes more restrictive.
}
\examples{

}
\author{
Eric R. Press
}
\concept{analytical functions}
