% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_functions.R
\name{clean_viewr}
\alias{clean_viewr}
\title{For an imported viewr object, run through the cleaning pipeline as desired}
\usage{
clean_viewr(
  obj_name,
  relabel_viewr_axes = TRUE,
  gather_tunnel_data = TRUE,
  trim_tunnel_outliers = TRUE,
  standardization_option = "rotate_tunnel",
  get_velocity = TRUE,
  select_x_percent = TRUE,
  separate_trajectories = TRUE,
  get_full_trajectories = TRUE,
  ...
)
}
\arguments{
\item{obj_name}{The viewr object to be processed}

\item{relabel_viewr_axes}{default TRUE,}

\item{gather_tunnel_data}{default TRUE,}

\item{trim_tunnel_outliers}{default TRUE,}

\item{standardization_option}{default "rotate_tunnel",}

\item{get_velocity}{default TRUE,}

\item{select_x_percent}{default TRUE,}

\item{separate_trajectories}{default TRUE,}

\item{get_full_trajectories}{default TRUE,}

\item{...}{Additional arugments passed to any of the corresponding functions}
}
\value{

}
\description{
For an imported viewr object, run through the cleaning pipeline as desired
}
