% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{clean_viewr_batch}
\alias{clean_viewr_batch}
\title{Batch clean viewr files}
\usage{
clean_viewr_batch(obj_list, ...)
}
\arguments{
\item{obj_list}{A list of viewr objects (i.e. a list of tibbles that each
have attribute \code{pathviewR_steps} that includes \code{"viewr"})}

\item{...}{Arguments to be passed in that specify how this function should
clean files.}
}
\value{
A list of viewr objects (tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}) that have been passed
through the corresponding cleaning functions.
}
\description{
For a list of viewr objects, run through the pipeline (from relabel axes
up through get full trajectories, as desired) via clean_viewr()
}
\details{
viewr objects should be in a list, e.g. the object generated by
\code{import_batch()}.

See \code{clean_viewr()} for details of how cleaning steps are handled
and/or refer to the corresponding cleaning functions themselves.
}
\seealso{
Other batch functions: 
\code{\link{bind_viewr_objects}()},
\code{\link{import_and_clean_batch}()},
\code{\link{import_batch}()}
}
\author{
Vikram B. Baliga
}
\concept{batch functions}
