% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Degrees (numeric)}
}
\value{

}
\description{
Convert degrees to radians
}
\seealso{
Other mathematical functions: 
\code{\link{find_curve_elbow}()},
\code{\link{get_3dcross_prod}()},
\code{\link{get_dist_point_line_2d}()},
\code{\link{get_dist_point_line_3d}()},
\code{\link{get_velocity}()},
\code{\link{rad2deg}()}
}
\concept{mathematical functions}
