% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{deg_2_rad}
\alias{deg_2_rad}
\title{Convert degrees to radians}
\usage{
deg_2_rad(deg)
}
\arguments{
\item{deg}{Degrees (a numeric of any length >= 1)}
}
\value{
The angle(s) in radians (as a numeric vector of the same length)
}
\description{
Convert degrees to radians
}
\examples{
## One input
deg_2_rad(90)

## Multiple inputs
deg_2_rad(c(5, 10, 15, 20))
}
\seealso{
Other mathematical functions: 
\code{\link{find_curve_elbow}()},
\code{\link{get_2d_angle}()},
\code{\link{get_3d_angle}()},
\code{\link{get_3d_cross_prod}()},
\code{\link{get_dist_point_line}()},
\code{\link{get_velocity}()},
\code{\link{rad_2_deg}()}
}
\author{
Vikram B. Baliga
}
\concept{mathematical functions}
