% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{get_3dcross_prod}
\alias{get_3dcross_prod}
\title{Compute the cross product of two 3D vectors}
\usage{
get_3dcross_prod(v1, v2)
}
\arguments{
\item{v1}{First vector, as c(x,y,z)}

\item{v2}{Second vector, as c(x,y,z)}
}
\value{

}
\description{
Compute the cross product of two 3D vectors
}
\seealso{
Other mathematical functions: 
\code{\link{deg2rad}()},
\code{\link{find_curve_elbow}()},
\code{\link{get_dist_point_line_2d}()},
\code{\link{get_dist_point_line_3d}()},
\code{\link{get_velocity}()},
\code{\link{rad2deg}()}
}
\concept{mathematical functions}
