% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{get_dist_point_line_3d}
\alias{get_dist_point_line_3d}
\title{Compute the distance between a point and a line in 3D space}
\usage{
get_dist_point_line_3d(point, line_coord1, line_coord2)
}
\arguments{
\item{point}{3D coordinates of the point as c(x, y, z)}

\item{line_coord1}{3D coordinates of one point on the line as c(x, y, z)}

\item{line_coord2}{3D coordinates of a second point on the line as c(x, y, z)}
}
\value{

}
\description{
Compute the distance between a point and a line in 3D space
}
\seealso{
Other mathematical functions: 
\code{\link{deg2rad}()},
\code{\link{find_curve_elbow}()},
\code{\link{get_3dcross_prod}()},
\code{\link{get_dist_point_line_2d}()},
\code{\link{get_velocity}()},
\code{\link{rad2deg}()}
}
\concept{mathematical functions}
