% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{get_traj_velocities}
\alias{get_traj_velocities}
\title{Recompute trajectory-specific velocities}
\usage{
get_traj_velocities(
  obj_name,
  time_col = "time_sec",
  length_col = "position_length",
  width_col = "position_width",
  height_col = "position_height",
  set_init_vel_zero = FALSE,
  velocity_min = NA,
  velocity_max = NA
)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewr_steps} that includes \code{"viewr"}}

\item{time_col}{Name of the column containing time}

\item{length_col}{Name of the column containing length dimension}

\item{width_col}{Name of the column containing width dimension}

\item{height_col}{Name of the column containing height dimension}

\item{set_init_vel_zero}{Should the first value be zero or can it be a
duplicate of the second velocity value? Defaults to FALSE.}

\item{velocity_min}{Should data below a certain velocity be filtered out of
the object? If so, enter a numeric. If not, keep NA.}

\item{velocity_max}{Should data above a certain velocity be filtered out of
the object? If so, enter a numeric. If not, keep NA.}
}
\value{
If \code{add_to_viewr} is \code{TRUE}, additional columns are
appended to the input viewr object. If \code{FALSE}, a standalone tibble is
created. Either way, an "instantaneous" velocity is computed as the
difference in position divided by the difference in time as each successive
row is encountered. Additionally, velocities along each of the three
position axes are computed and provided as additional columns.
}
\description{
Recompute trajectory-specific velocities
}
\details{
Instantaneous velocity is not truly "instantaneous" but rather is
approximated as the change in distance divided by change in time from one
observation (row) to the previous observation (row). Each component of
velocity is computed (i.e. per axis) along with the overall velocity of
the subject.
}
\seealso{
Other mathematical functions: 
\code{\link{calc_min_dist_v}()},
\code{\link{deg_2_rad}()},
\code{\link{find_curve_elbow}()},
\code{\link{get_2d_angle}()},
\code{\link{get_3d_angle}()},
\code{\link{get_3d_cross_prod}()},
\code{\link{get_dist_point_line}()},
\code{\link{get_velocity}()},
\code{\link{rad_2_deg}()}
}
\author{
Vikram B. Baliga
}
\concept{mathematical functions}
