% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{get_velocity}
\alias{get_velocity}
\title{Get instantaneous velocity for subjects}
\usage{
get_velocity(
  obj_name,
  time_col = "time_sec",
  length_col = "position_length",
  width_col = "position_width",
  height_col = "position_height",
  add_to_viewr = TRUE,
  velocity_min = NA,
  velocity_max = NA,
  ...
)
}
\arguments{
\item{obj_name}{Input viewr object}

\item{time_col}{Name of the column containing time}

\item{length_col}{Name of the column containing length dimension}

\item{width_col}{Name of the column containing width dimension}

\item{height_col}{Name of the column containing height dimension}

\item{add_to_viewr}{Default TRUE; should velocity data be added as new
columns or should this function create a new simpler object?}

\item{velocity_min}{Should data below a certain velocity be filtered out? If
so, enter a numeric. If not, keep NA.}

\item{velocity_max}{Should data above a certain velocity be filtered out? If
so, enter a numeric. If not, keep NA.}

\item{...}{Additional arguments.}
}
\value{

}
\description{
Get instantaneous velocity for subjects
}
\seealso{
Other mathematical functions: 
\code{\link{deg2rad}()},
\code{\link{find_curve_elbow}()},
\code{\link{get_3dcross_prod}()},
\code{\link{get_dist_point_line_2d}()},
\code{\link{get_dist_point_line_3d}()},
\code{\link{rad2deg}()}
}
\concept{mathematical functions}
