% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{get_vis_angle}
\alias{get_vis_angle}
\title{Estimate visual angles from a subject's perspective in an experimental tunnel}
\usage{
get_vis_angle(obj_name)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with
attributes \code{pathviewR_steps} that include \code{"viewr"}
and \code{min_dist_calculated}.}
}
\value{
A tibble or data.frame with added variables for
\code{vis_angle_pos_rad}, \code{vis_angle_pos_deg},
\code{vis_angle_neg_rad}, \code{vos_angle_neg_deg},
\code{vis_angle_end_rad}, and \code{vis_angle_end_deg}.
}
\description{
Estimate visual angles from a subject's perspective in an experimental tunnel
}
\details{
\code{get_vis_angle()} assumes the subject's gaze is fixed at the
point on the either side of the tunnel that minimizes the distance to
visual stimuli and therefore maximizes visual angles. Angles are reported
in radians/cycle (\code{vis_angle_pos_rad}) or degrees/cycle
(\code{vis_angle_pos_deg}).
}
\seealso{
Other visual perception functions: 
\code{\link{get_sf}()}
}
\author{
Eric R. Press
}
\concept{visual perception functions}
