% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_functions.R
\name{import_and_clean_viewr}
\alias{import_and_clean_viewr}
\title{Import a file and run through as many cleaning steps as desired.}
\usage{
import_and_clean_viewr(
  file_name,
  file_id = NA,
  relabel_viewr_axes = TRUE,
  gather_tunnel_data = TRUE,
  trim_tunnel_outliers = TRUE,
  standardization_option = "rotate_tunnel",
  get_velocity = TRUE,
  select_x_percent = TRUE,
  rename_viewr_characters = FALSE,
  separate_trajectories = TRUE,
  get_full_trajectories = TRUE,
  ...
)
}
\arguments{
\item{file_name}{Target file}

\item{file_id}{Optional}

\item{relabel_viewr_axes}{default TRUE,}

\item{gather_tunnel_data}{default TRUE,}

\item{trim_tunnel_outliers}{default TRUE,}

\item{standardization_option}{default TRUE,}

\item{get_velocity}{default TRUE,}

\item{select_x_percent}{default TRUE,}

\item{rename_viewr_characters}{default FALSE,}

\item{separate_trajectories}{default TRUE,}

\item{get_full_trajectories}{default TRUE,}

\item{...}{Additional arguments passed to the corresponding functions.}
}
\value{

}
\description{
Import a file and run through as many cleaning steps as desired.
}
