% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{import_batch}
\alias{import_batch}
\title{Batch import of files for either Motive or Flydra (but not a mix of both)}
\usage{
import_batch(
  file_path_list,
  import_method = c("flydra", "motive"),
  file_id = NA,
  subject_name = NULL,
  frame_rate = NULL,
  simplify_marker_naming = TRUE,
  import_messaging = FALSE,
  ...
)
}
\arguments{
\item{file_path_list}{A list of file paths}

\item{import_method}{Either "flydra" or "motive"}

\item{file_id}{Optional}

\item{subject_name}{For Flydra, the assigned subject name}

\item{frame_rate}{For Flydra, the assigned frame rate}

\item{simplify_marker_naming}{default TRUE; for Motive, whether marker naming
should be simplified}

\item{import_messaging}{default FALSE; should this function report each time
a file has been imported?}

\item{...}{Additional arguments (may remove this if needed)}
}
\value{

}
\description{
Batch import of files for either Motive or Flydra (but not a mix of both)
}
\seealso{
Other data import functions: 
\code{\link{import_and_clean_batch}()},
\code{\link{read_flydra_data}()},
\code{\link{read_motive_csv}()}

Other batch functions: 
\code{\link{bind_viewr_objects}()},
\code{\link{clean_viewr_batch}()},
\code{\link{import_and_clean_batch}()}
}
\concept{batch functions}
\concept{data import functions}
