% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathviewR.R
\docType{package}
\name{pathviewR-package}
\alias{pathviewR}
\alias{pathviewR-package}
\title{pathviewR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
Description of package TBD
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
