% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{Radians (numeric)}
}
\value{

}
\description{
Convert radians to degrees
}
\seealso{
Other mathematical functions: 
\code{\link{deg2rad}()},
\code{\link{find_curve_elbow}()},
\code{\link{get_3dcross_prod}()},
\code{\link{get_dist_point_line_2d}()},
\code{\link{get_dist_point_line_3d}()},
\code{\link{get_velocity}()}
}
\concept{mathematical functions}
