% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{rad_2_deg}
\alias{rad_2_deg}
\title{Convert radians to degrees}
\usage{
rad_2_deg(rad)
}
\arguments{
\item{rad}{Radians (a numeric of any length >= 1)}
}
\value{
The angle(s) in degrees (as a numeric vector of the same length)
}
\description{
Convert radians to degrees
}
\examples{
## One input
rad_2_deg(pi/2)

## Multiple inputs
rad_2_deg(c(pi / 2, pi, 2 * pi))
}
\seealso{
Other mathematical functions: 
\code{\link{deg_2_rad}()},
\code{\link{find_curve_elbow}()},
\code{\link{get_3d_cross_prod}()},
\code{\link{get_dist_point_line_2d}()},
\code{\link{get_dist_point_line_3d}()},
\code{\link{get_velocity}()}
}
\author{
Vikram B. Baliga
}
\concept{mathematical functions}
