% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import_functions.R
\name{read_flydra_data}
\alias{read_flydra_data}
\title{Import data from a MAT file exported from Flydra software}
\usage{
read_flydra_data(mat_file, file_id = NA, subject_name, frame_rate = 100, ...)
}
\arguments{
\item{mat_file}{The MAT file from Flydra}

\item{file_id}{(Optional) identifier for this file. If not supplied, this
defaults to \code{basename(file_name)}.}

\item{subject_name}{Name that will be assigned to the subject}

\item{frame_rate}{The capture frame rate of the session}

\item{...}{Additional arguments}
}
\value{

}
\description{
Import data from a MAT file exported from Flydra software
}
\seealso{
\code{\link{read_motive_csv}} for importing Motive data

Other data import functions: 
\code{\link{import_and_clean_batch}()},
\code{\link{import_batch}()},
\code{\link{read_motive_csv}()}
}
\concept{data import functions}
