% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{redefine_tunnel_center}
\alias{redefine_tunnel_center}
\title{"Center" the tunnel data, i.e. translation but no rotation}
\usage{
redefine_tunnel_center(
  obj_name,
  axes = c("position_length", "position_width", "position_height"),
  length_method = c("original", "middle", "median", "user-defined"),
  width_method = c("original", "middle", "median", "user-defined"),
  height_method = c("original", "middle", "median", "user-defined"),
  length_zero = NA,
  width_zero = NA,
  height_zero = NA,
  ...
)
}
\arguments{
\item{obj_name}{The target viewr object}

\item{axes}{Names of axes to be centered}

\item{length_method}{Method for length}

\item{width_method}{Method for width}

\item{height_method}{Method for height}

\item{length_zero}{User-defined value}

\item{width_zero}{User-defined value}

\item{height_zero}{User-defined value}

\item{...}{Additional arguments}
}
\value{

}
\description{
"Center" the tunnel data, i.e. translation but no rotation
}
\seealso{
Other data cleaning functions: 
\code{\link{gather_tunnel_data}()},
\code{\link{get_full_trajectories}()},
\code{\link{quick_separate_trajectories}()},
\code{\link{relabel_viewr_axes}()},
\code{\link{rename_viewr_characters}()},
\code{\link{rotate_tunnel}()},
\code{\link{select_x_percent}()},
\code{\link{separate_trajectories}()},
\code{\link{standardize_tunnel}()},
\code{\link{trim_tunnel_outliers}()},
\code{\link{visualize_frame_gap_choice}()}

Other tunnel standardization functions: 
\code{\link{rotate_tunnel}()},
\code{\link{standardize_tunnel}()}
}
\concept{data cleaning functions}
\concept{tunnel standardization functions}
