% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{relabel_motiv_axes}
\alias{relabel_motiv_axes}
\title{Relabel the dimensions as length, width, and height}
\usage{
relabel_motiv_axes(
  obj_name,
  tunnel_length = "_Z",
  tunnel_width = "_X",
  tunnel_height = "_Y",
  ...
)
}
\arguments{
\item{obj_name}{An object of class \code{motiv}}

\item{tunnel_length}{The dimension that corresponds to tunnel length. Set to
\code{tunnel_length = "_Z"} by default. Argument should contain a character
vector with a leading underscore (see Details)}

\item{tunnel_width}{The dimension that corresponds to tunnel width. Follows
the same conventions as \code{tunnel_length} and defaults to
\code{tunnel_length = "_X"}}

\item{tunnel_height}{The dimension that corresponds to tunnel height. Follows
the same conventions as \code{tunnel_length} and defaults to
\code{tunnel_length = "_Y"}}

\item{...}{Additional arguments to be passed to \code{read_motiv_csv()}.}
}
\value{
An object of class \code{motiv} with variables that have been
renamed.
}
\description{
Axes are commonly labeled as "X", "Y", and "Z" in recording software yet
you may desire to relabel these as "length", "width", and "height".
\code{relabel_motiv_axes()} is a function that takes a \code{motiv} object
and allows the user to rename the variables.
}
\details{
Each argument must have a leading underscore ("_") and each
argument must have an entry. E.g. tunnel_length = "_Y" will replace all
instances of _Y with _length in the names of variables.
}
\examples{

library(pathviewR)

## Import the july 29 example data included in the package
jul_29 <-
  read_motiv_csv(system.file("extdata", "july-29_group-I_16-20.csv",
                             package = 'pathviewR'))

## Names of variables are labeled with _X, _Y, _Z, which we'd like to rename
names(jul_29)

## Now use relabel_motiv_axes() to rename these variables using _length,
_width, and _height instead
jul_29_relabeled <-
  relabel_motiv_axes(jul_29,
                     tunnel_length = "_Z",
                     tunnel_width = "_X",
                     tunnel_height = "_Y")

## See the result
names(jul_29_relabeled)

}
\seealso{
\code{\link{read_motiv_csv}}
}
\author{
Vikram B. Baliga
}
\concept{utility functions}
