% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{relabel_viewr_axes}
\alias{relabel_viewr_axes}
\title{Relabel the dimensions as length, width, and height}
\usage{
relabel_viewr_axes(
  obj_name,
  tunnel_length = "_z",
  tunnel_width = "_x",
  tunnel_height = "_y",
  real = "_w",
  ...
)
}
\arguments{
\item{obj_name}{A tibble or data.frame with attribute \code{viewr}}

\item{tunnel_length}{The dimension that corresponds to tunnel length. Set to
\code{tunnel_length = "_z"} by default. Argument should contain a character
vector with a leading underscore (see Details)}

\item{tunnel_width}{The dimension that corresponds to tunnel width. Follows
the same conventions as \code{tunnel_length} and defaults to
\code{tunnel_length = "_x"}}

\item{tunnel_height}{The dimension that corresponds to tunnel height. Follows
the same conventions as \code{tunnel_length} and defaults to
\code{tunnel_length = "_y"}}

\item{real}{The dimension that corresponds to the "real" parameter in
quaternion notation (for data with "rotation" values). Follows the same
conventions as \code{tunnel_length} and defaults to \code{real = "_w"}}

\item{...}{Additional arguments to be passed to \code{read_motive_csv()}.}
}
\value{
A tibble or data.frame with variables that have been renamed.
}
\description{
Axes are commonly labeled as "x", "y", and "z" in recording software yet
you may desire to relabel these as "length", "width", and "height".
\code{relabel_viewr_axes()} is a function that takes a \code{viewr} object
and allows the user to rename the variables.
}
\details{
Each argument must have a leading underscore ("_") and each
argument must have an entry. E.g. tunnel_length = "_Y" will replace all
instances of _Y with _length in the names of variables.
}
\examples{

library(pathviewR)

## Import the july 29 example data included in the package
jul_29 <-
  read_motive_csv(system.file("extdata", "july-29_group-I_16-20.csv",
                             package = 'pathviewR'))

## Names of variables are labeled with _x, _y, _z, which we'd like to rename
names(jul_29)

## Now use relabel_viewr_axes() to rename these variables using _length,
_width, and _height instead
jul_29_relabeled <-
  relabel_viewr_axes(jul_29,
                     tunnel_length = "_z",
                     tunnel_width = "_x",
                     tunnel_height = "_y",
                     real = "_w")

## See the result
names(jul_29_relabeled)

}
\seealso{
\code{\link{read_motive_csv}}
}
\author{
Vikram B. Baliga
}
\concept{utility functions}
