% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{remove_duplicate_frames}
\alias{remove_duplicate_frames}
\title{Remove any duplicates or aliased frames within trajectories}
\usage{
remove_duplicate_frames(obj_name)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewr_steps} that includes \code{"viewr"}}
}
\value{
A viewr object (tibble or data.frame with attribute
\code{pathviewr_steps}.
}
\description{
Remove any duplicates or aliased frames within trajectories
}
\details{
The separate_trajectories() and get_full_trajectories() must be
run prior to use.
}
\seealso{
Other utility functions: 
\code{\link{clean_by_span}()},
\code{\link{insert_treatments}()},
\code{\link{remove_vel_anomalies}()},
\code{\link{set_traj_frametime}()}
}
\author{
Vikram B. Baliga
}
\concept{utility functions}
