% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rm_by_trajnum}
\alias{rm_by_trajnum}
\title{Remove subjects by trajectory number}
\usage{
rm_by_trajnum(obj_name, trajnum = 5, stim1, stim2, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Remove subjects by trajectory number
}
