% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{section_tunnel_by}
\alias{section_tunnel_by}
\title{Bin data along a specified axis}
\usage{
section_tunnel_by(obj_name, axis = "position_length", number_of_sections = 20)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}

\item{axis}{Chosen axis, must match name of column exactly}

\item{number_of_sections}{Total number of sections}
}
\value{
A new column added to the input data object called \code{section_id},
which is an ordered factor that indicates grouping.
}
\description{
Chop data into X sections (of equal size) along a specified axis
}
\details{
The idea is to bin the data along a specified axis, generally
\code{position_length}.
}
\author{
Vikram B. Baliga
}
