% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{separate_trajectories}
\alias{separate_trajectories}
\title{Separate rows of data into separately labeled trajectories.}
\usage{
separate_trajectories(
  obj_name,
  max_frame_gap = 1,
  frame_rate_proportion = 0.1,
  frame_gap_messaging = FALSE,
  frame_gap_plotting = FALSE,
  ...
)
}
\arguments{
\item{obj_name}{Target object}

\item{max_frame_gap}{If known, a numeric value to use}

\item{frame_rate_proportion}{If autodetect, proportion of frame rate as
reference}

\item{frame_gap_messaging}{Should frame gaps be reported in the console?}

\item{frame_gap_plotting}{Should frame gap diagnostic plots be shown?}

\item{...}{Additional arguments}
}
\value{

}
\description{
Separate rows of data into separately labeled trajectories.
}
