% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{visualize_frame_gap_choice}
\alias{visualize_frame_gap_choice}
\title{Run separate_trajectories() with many different frame gaps to help determine
what value to use}
\usage{
visualize_frame_gap_choice(obj_name, loops = 20, ...)
}
\arguments{
\item{obj_name}{Input viewr object}

\item{loops}{How many total frame gap entries to consider}

\item{...}{Additional arguments}
}
\value{

}
\description{
Run separate_trajectories() with many different frame gaps to help determine
what value to use
}
