% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{bind_viewr_objects}
\alias{bind_viewr_objects}
\title{Combine multiple viewr objects into a single viwer object}
\usage{
bind_viewr_objects(obj_list)
}
\arguments{
\item{obj_list}{A list of viewr objects}
}
\value{

}
\description{
Combine multiple viewr objects into a single viwer object
}
\seealso{
Other batch functions: 
\code{\link{clean_viewr_batch}()},
\code{\link{import_and_clean_batch}()},
\code{\link{import_batch}()}
}
\concept{batch functions}
