% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{calc_sf_box}
\alias{calc_sf_box}
\title{Estimate spatial frequency of visual stimuli in rectangular tunnel, i.e. box}
\usage{
calc_sf_box(obj_name)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}
}
\value{
A tibble or data.frame with added variables \code{min_dist_pos} and
\code{min_dist_neg} describing the minimum distance to the tunnel walls as
well as \code{s_freq_pos} and \code{s_freq_neg} reported in cycles per 1deg
of visual angle.
}
\description{
Based on animal head positions in a rectangular tunnel i.e. box,
\code{calc_sf_box()} calculates how the animal perceives the spatial
frequency of lateral visual stimuli as modulated by distance to the stimulus.
}
\details{
\code{calc_sf_box} assumes fixed gaze at the point on the
either side of the tunnel that minimizes the distance to visual stimuli and
thereby minimizes spatial frequencies.
}
\seealso{
Other visual perception functions: 
\code{\link{calc_sf_V}()},
\code{\link{calc_vis_angle_V}()},
\code{\link{calc_vis_angle_box}()}
}
\author{
Eric R. Press
}
\concept{visual perception functions}
