% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{calc_vis_angle_box}
\alias{calc_vis_angle_box}
\title{Estimate visual angles in a box-shaped tunnel}
\usage{
calc_vis_angle_box(obj_name)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"} and
\code{treatments_added}.}
}
\value{
A tibble or data.frame with added variables for \code{min_dist_pos},
\code{min_dist_neg}, \code{vis_angle_pos_rad}, \code{vis_angle_neg_rad},
\code{vis_angle_pos_deg}, and \code{vis_angle_neg_deg}.
}
\description{
Based on rigid body, i.e animal head positions in a b-shaped tunnel,
\code{calc_vis_angle_box()} calculates the visual angles created by lateral
visual stimuli.
}
\details{
\code{cal_vis_angle_box} assumes fixed gaze at the point on the
either side of the tunnel that minimizes the distance to visual stimuli and
thereby maximizes visual angles. All length measurements reported in
meters.
}
\examples{

}
\seealso{
Other visual perception functions: 
\code{\link{calc_vis_angle_V}()}
}
\author{
Eric R. Press
}
\concept{visual perception functions}
