% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one_functions.R
\name{clean_viewr}
\alias{clean_viewr}
\title{All-in-one function to clean imported objects}
\usage{
clean_viewr(
  obj_name,
  relabel_viewr_axes = TRUE,
  gather_tunnel_data = TRUE,
  trim_tunnel_outliers = TRUE,
  standardization_option = "rotate_tunnel",
  get_velocity = TRUE,
  select_x_percent = TRUE,
  rename_viewr_characters = FALSE,
  separate_trajectories = TRUE,
  get_full_trajectories = TRUE,
  ...
)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}

\item{relabel_viewr_axes}{default TRUE, should axes be relabeled?}

\item{gather_tunnel_data}{default TRUE, should tunnel data be gathered?}

\item{trim_tunnel_outliers}{default TRUE, outliers be trimmed?}

\item{standardization_option}{default "rotate_tunnel"; which standardization
option should be used? See Details for more.}

\item{get_velocity}{default TRUE, should velocity be computed?}

\item{select_x_percent}{default TRUE, should a region of interest be
selected?}

\item{rename_viewr_characters}{default FALSE, should subjects be renamed?}

\item{separate_trajectories}{default TRUE, should trajectories be defined?}

\item{get_full_trajectories}{default TRUE, should only full trajectories be
retained.}

\item{...}{Additional arguments passed to any of the corresponding functions}
}
\value{
A viewr object (tibble or data.frame with attribute \code{pathviewR_steps} that includes \code{"viewr"}) that has passed through several \code{pathviewR} functions as desired by the user, resulting in data that have been cleaned and ready for analyses.
}
\description{
For an imported viewr object, run through the cleaning pipeline as desired
}
\details{
Each argument corresponds to a standalone function in
\code{pathviewR}. E.g. the parameter \code{relabel_viewr_axes} allows a
user to choose whether \code{pathviewR::relabel_viewr_axes()} is run
internally. Should the user desire to use any non-default parameter values
for any functions included here, they should be supplied to this function
as additional arguments formatted exactly as they would appear in their
corresponding function(s). E.g. if the "autodetect" feature in
\code{pathviewR::separate_trajectories()} is desired, add an argument
\code{max_frame_gap = "autodetect"} to the arguments supplied to this
function.
}
\seealso{
Other all in one functions: 
\code{\link{import_and_clean_viewr}()}
}
\author{
Vikram B. Baliga
}
\concept{all in one functions}
