% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{clean_viewr_batch}
\alias{clean_viewr_batch}
\title{For a list of viewr files, run through the pipeline (from relabel axes
up through get full trajectories, as desired) via clean_viewr()}
\usage{
clean_viewr_batch(obj_list, ...)
}
\arguments{
\item{obj_list}{A list of viewr files (i.e. a list of tibbles that each have
attribute \code{pathviewR_steps} that includes \code{"viewr"})}

\item{...}{Arguments to be passed in that specify how this function should
clean files.}
}
\value{

}
\description{
For a list of viewr files, run through the pipeline (from relabel axes
up through get full trajectories, as desired) via clean_viewr()
}
\seealso{
Other batch functions: 
\code{\link{bind_viewr_objects}()},
\code{\link{import_and_clean_batch}()},
\code{\link{import_batch}()}
}
\author{
Vikram B. Baliga
}
\concept{batch functions}
