% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_functions.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{Degrees (a numeric of any length >= 1)}
}
\value{
The angle(s) in radians (as a numeric vector of the same length)
}
\description{
Convert degrees to radians
}
\examples{
## One input
deg2rad(90)

## Multiple inputs
deg2rad(c(5, 10, 15, 20))
}
\seealso{
Other mathematical functions: 
\code{\link{find_curve_elbow}()},
\code{\link{get_3d_cross_prod}()},
\code{\link{get_dist_point_line_2d}()},
\code{\link{get_dist_point_line_3d}()},
\code{\link{get_velocity}()},
\code{\link{rad2deg}()}
}
\author{
Vikram B. Baliga
}
\concept{mathematical functions}
