% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_full_trajectories}
\alias{get_full_trajectories}
\title{Specify a minimum span of the (selected) tunnel and then keep trajectories
that are wider than that span and go from one end to the other}
\usage{
get_full_trajectories(obj_name, span = 0.8, ...)
}
\arguments{
\item{obj_name}{Target viewr object}

\item{span}{Span to use; must range from 0 to 1}

\item{...}{Additional arguments}
}
\value{

}
\description{
Specify a minimum span of the (selected) tunnel and then keep trajectories
that are wider than that span and go from one end to the other
}
