% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_header_viewr}
\alias{get_header_viewr}
\title{Extract header info from imported viewr object}
\usage{
get_header_viewr(obj_name, ...)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}
\code{pathviewR_steps}}

\item{...}{Additional arguments that may be passed to other \code{pathviewR}
functions}
}
\value{
The value of the \code{header} attribute, or NULL if no exact match
is found and no or more than one partial match is found.
}
\description{
A function to quickly return the information stored in the header of the
original data file imported via \code{pathviewR}'s \code{read_} functions.
}
\examples{
library(pathviewR)

## Import the july 29 example data included in the package
jul_29 <-
  read_motive_csv(system.file("extdata", "july-29_group-I_16-20.csv",
                             package = 'pathviewR'))

## Now display the Header information
get_header_viewr(jul_29)
}
\author{
Vikram B. Baliga
}
\concept{metadata handling functions}
