% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{import_and_clean_batch}
\alias{import_and_clean_batch}
\title{Like clean viewr batch, but with import as the first step too}
\usage{
import_and_clean_batch(
  file_path_list,
  import_method = c("flydra", "motive"),
  file_id = NA,
  subject_name = NULL,
  frame_rate = NULL,
  simplify_marker_naming = TRUE,
  import_messaging = FALSE,
  ...
)
}
\arguments{
\item{file_path_list}{A list of file paths}

\item{import_method}{Either "flydra" or "motive"}

\item{file_id}{Optional}

\item{subject_name}{For Flydra, the subject name applied to all files}

\item{frame_rate}{For Flydra, the frame rate applied to all files
simplified?}

\item{import_messaging}{Should this function report each time a file has been
processed?}

\item{...}{Additional arguments to specify how data should be cleaned.}
}
\value{

}
\description{
Like clean viewr batch, but with import as the first step too
}
