% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{import_and_clean_batch}
\alias{import_and_clean_batch}
\title{Batch import and clean files}
\usage{
import_and_clean_batch(
  file_path_list,
  import_method = c("flydra", "motive"),
  file_id = NA,
  subject_name = NULL,
  frame_rate = NULL,
  simplify_marker_naming = TRUE,
  import_messaging = FALSE,
  ...
)
}
\arguments{
\item{file_path_list}{A list of file paths leading to files to be imported.}

\item{import_method}{Either "flydra" or "motive"}

\item{file_id}{(Optional) identifier for this file. If not supplied, this
defaults to \code{basename(file_name)}.}

\item{subject_name}{For Flydra, the subject name applied to all files}

\item{frame_rate}{For Flydra, the frame rate applied to all files}

\item{simplify_marker_naming}{For Motive, if Markers are encountered, should
they be renamed from "Subject:marker" to "marker"? Defaults to TRUE}

\item{import_messaging}{Should this function report each time a file has been
processed?}

\item{...}{Additional arguments to specify how data should be cleaned.}
}
\value{
A list of viewr objects (tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}) that have been passed
through the corresponding cleaning functions.
}
\description{
Like \code{clean_viewr_batch()}, but with import as the first step too
}
\details{
viewr objects should be in a list, e.g. the object generated by
\code{import_batch()}.

See \code{clean_viewr()} for details of how cleaning steps are handled
and/or refer to the corresponding cleaning functions themselves.
}
\examples{
## Since we only have one example file of each type provided
## in pathviewR, we will simply import the same example multiple
## times to simulate batch importing. Replace the contents of
## the following list with your own list of files to be imported.

## Make a list of the same example file 3x
import_list <-
  c(rep(
    system.file("extdata", "pathviewR_motive_example_data.csv",
                package = 'pathviewR'),
    3
  ))

## Batch import
motive_batch_imports <-
  import_batch(import_list,
               import_method = "motive",
               import_messaging = TRUE)

## Batch cleaning of these imported files
## via clean_viewr_batch()
motive_batch_cleaned <-
  clean_viewr_batch(
    motive_batch_imports,
    desired_percent = 50,
    max_frame_gap = "autodetect",
    span = 0.95
  )

## Alternatively, use import_and_clean_batch() to
## combine import with cleaning on a batch of files
motive_batch_import_and_clean <-
  import_and_clean_batch(
    import_list,
    import_method = "motive",
    import_messaging = TRUE,
    motive_batch_imports,
    desired_percent = 50,
    max_frame_gap = "autodetect",
    span = 0.95
  )

## Each of these lists of objects can be bound into
## one viewr object (i.e. one tibble) via
## bind_viewr_objects()
motive_bound_one <-
  bind_viewr_objects(motive_batch_cleaned)

motive_bound_two <-
  bind_viewr_objects(motive_batch_import_and_clean)

## Either route results in the same object ultimately:
identical(motive_bound_one, motive_bound_two)
}
\seealso{
Other data import functions: 
\code{\link{as_viewr}()},
\code{\link{import_batch}()},
\code{\link{read_flydra_mat}()},
\code{\link{read_motive_csv}()}

Other batch functions: 
\code{\link{bind_viewr_objects}()},
\code{\link{clean_viewr_batch}()},
\code{\link{import_batch}()}
}
\author{
Vikram B. Baliga
}
\concept{batch functions}
\concept{data import functions}
