% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_functions.R
\name{import_and_clean_batch}
\alias{import_and_clean_batch}
\title{Like clean viewr batch, but with import as the first step too}
\usage{
import_and_clean_batch(
  file_path_list,
  import_method = c("flydra", "motive"),
  file_id = NA,
  subject_name = NULL,
  frame_rate = NULL,
  simplify_marker_naming = TRUE,
  import_messaging = FALSE,
  ...
)
}
\arguments{
\item{file_path_list}{A list of file paths}

\item{import_method}{Either "flydra" or "motive"}

\item{file_id}{(Optional) identifier for this file. If not supplied, this
defaults to \code{basename(file_name)}.}

\item{subject_name}{For Flydra, the subject name applied to all files}

\item{frame_rate}{For Flydra, the frame rate applied to all files}

\item{simplify_marker_naming}{For Motive, if Markers are encountered, should
they be renamed from "Subject:marker" to "marker"? Defaults to TRUE}

\item{import_messaging}{Should this function report each time a file has been
processed?}

\item{...}{Additional arguments to specify how data should be cleaned.}
}
\value{

}
\description{
Like clean viewr batch, but with import as the first step too
}
\seealso{
Other data import functions: 
\code{\link{import_batch}()},
\code{\link{read_flydra_mat}()},
\code{\link{read_motive_csv}()}

Other batch functions: 
\code{\link{bind_viewr_objects}()},
\code{\link{clean_viewr_batch}()},
\code{\link{import_batch}()}
}
\concept{batch functions}
\concept{data import functions}
