% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{insert_treatments}
\alias{insert_treatments}
\title{Inserts treatment and experiment information}
\usage{
insert_treatments(
  obj_name,
  vertex_height = NULL,
  vertex_angle = NULL,
  pos_wall = NULL,
  neg_wall = NULL,
  front_wall = NULL,
  stim_param_pos = NULL,
  stim_param_neg = NULL,
  stim_param_front = NULL,
  treatment = NULL
)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}

\item{vertex_height}{If using a V-shaped chamber, this is the height of the
vertex. This value should be negative and represents the distance between the
origin (0,0,0) and the vertex.}

\item{vertex_angle}{If using a V-shaped chamber, this is the angle (in degrees)
subtended by a vertical axis and the sides of the V-shaped chamber.
Equivalent to half the angle of the "V". \code{vertex_angle} defaults to 45.}

\item{pos_wall}{If using a box-shaped chamber, this is the distance in meters between the origin (0,0,0) and the
positive wall.}

\item{neg_wall}{If using a box-shaped chamber, this is the distance in meters between the origin (0,0,0) and the
negative wall.}

\item{front_wall}{If using a box-shaped chamber, this is the distance in meters between the origin (0,0,0) and the
front wall.}

\item{stim_param_pos}{The width of the sine-wave grating stimulus displayed on the positive wall of the chamber.}

\item{stim_param_neg}{The width of the sine-wave grating stimulus displayed on the negative wall of the chamber.}

\item{stim_param_front}{If using a box-shaped chamber, the width of the sine-wave grating displayed on the front wall.}

\item{treatment}{The treatment name assigned to all subjects in the viewr
object. Currently only able to accept a single treatment per viewr object.}
}
\value{
A viewr object (tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"treatments added"}). Depending on the chamber configuration, a tibble or data.frame with added variables for
\code{vertex_height}, \code{vertex_angle}, \code{pos_wall}, \code{neg_wall}, \code{front_wall}, and
\code{treatment}. This experiment information is also stored in the viewr object's metadata
}
\description{
Adds information about treatment and experimental set up to viewr objects for
analysis in other pathviewR functions
}
\details{
All length measurements reported in meters.
}
\examples{

}
\author{
Eric R. Press
}
