% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{insert_treatments}
\alias{insert_treatments}
\title{Inserts treatment and experiment information}
\usage{
insert_treatments(
  obj_name,
  tunnel_config = "box",
  perch_2_vertex = NULL,
  vertex_angle = NULL,
  tunnel_width = NULL,
  tunnel_length = NULL,
  stim_param_lat_pos = NULL,
  stim_param_lat_neg = NULL,
  stim_param_end_pos = NULL,
  stim_param_end_neg = NULL,
  treatment = NULL
)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}

\item{tunnel_config}{The configuration of the experimental tunnel.
Currently, pathviewR supports rectangular "box" and V-shaped tunnel
configurations.}

\item{perch_2_vertex}{If using a V-shaped tunnel, this is the vertical
distance between the vertex and the height of the perches. If the tunnel does
not have perches, insert the vertical distance between the vertex and the
height of the origin (0,0,0).}

\item{vertex_angle}{If using a V-shaped tunnel, the angle of the vertex (in
degrees) \code{vertex_angle} defaults to 90.}

\item{tunnel_width}{If using a box-shaped tunnel, the width of the tunnel.}

\item{tunnel_length}{The length of the tunnel.}

\item{stim_param_lat_pos}{The size of the stimulus on the lateral positive
wall of the tunnel. Eg. for 10cm wide gratings,
\code{stim_param_lat_pos} = 0.1.}

\item{stim_param_lat_neg}{The size of the stimulus on the lateral negative
wall of the tunnel..}

\item{stim_param_end_pos}{The size of the stimulus on the end positive
wall of the tunnel.}

\item{stim_param_end_neg}{The size of the stimulus on the end negative
wall of the tunnel.}

\item{treatment}{The name of the treatment assigned to all rows of the viewr
object. Currently only able to accept a single treatment per viewr data
object.}
}
\value{
A viewr object (tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"treatments added"}). Depending
on the argument \code{tunnel_config}, the viewr object also includes
columns storing the values of the supplied arguments. This experimental
information is also stored in the viewr object's metadata
}
\description{
Adds information about treatment and experimental set up to viewr objects for
analysis in other pathviewR functions
}
\details{
All length measurements reported in meters.
}
\author{
Eric R. Press
}
\concept{utility functions}
