% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathviewR.R
\docType{package}
\name{pathviewR-package}
\alias{pathviewR}
\alias{pathviewR-package}
\title{pathviewR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
Description of package TBD
}
\author{
\strong{Maintainer}: Vikram B. Baliga \email{vbaliga87@gmail.com} (\href{https://orcid.org/0000-0002-9367-8974}{ORCID})

Authors:
\itemize{
  \item Melissa S. Armstrong \email{FILLME@EMAIL.DOMAIN} (\href{https://orcid.org/XXXX-XXXX-XXXX-XXXX}{ORCID})
  \item Eric R. Press \email{epress12@gmail.com} (\href{https://orcid.org/0000-0002-1944-3755}{ORCID})
}

}
\keyword{internal}
