% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{quick_separate_trajectories}
\alias{quick_separate_trajectories}
\title{Quick version of separate_trajectories}
\usage{
quick_separate_trajectories(obj_name, max_frame_gap = 1, ...)
}
\arguments{
\item{obj_name}{Target viwer object}

\item{max_frame_gap}{Must be numeric}

\item{...}{Additional arguments}
}
\value{

}
\description{
Quick version of separate_trajectories
}
