% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{quick_separate_trajectories}
\alias{quick_separate_trajectories}
\title{Quick version of separate_trajectories}
\usage{
quick_separate_trajectories(obj_name, max_frame_gap = 1, ...)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}

\item{max_frame_gap}{Must be numeric}

\item{...}{Additional arguments}
}
\value{

}
\description{
Quick version of separate_trajectories
}
\seealso{
Other data cleaning functions: 
\code{\link{gather_tunnel_data}()},
\code{\link{get_full_trajectories}()},
\code{\link{redefine_tunnel_center}()},
\code{\link{relabel_viewr_axes}()},
\code{\link{rename_viewr_characters}()},
\code{\link{rotate_tunnel}()},
\code{\link{select_x_percent}()},
\code{\link{separate_trajectories}()},
\code{\link{standardize_tunnel}()},
\code{\link{trim_tunnel_outliers}()},
\code{\link{visualize_frame_gap_choice}()}

Other functions that define or clean trajectories: 
\code{\link{get_full_trajectories}()},
\code{\link{separate_trajectories}()},
\code{\link{visualize_frame_gap_choice}()}
}
\author{
Vikram B. Baliga
}
\concept{data cleaning functions}
\concept{functions that define or clean trajectories}
