% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{redefine_tunnel_center}
\alias{redefine_tunnel_center}
\title{"Center" the tunnel data, i.e. translation but no rotation}
\usage{
redefine_tunnel_center(
  obj_name,
  axes = c("position_length", "position_width", "position_height"),
  length_method = c("original", "middle", "median", "user-defined"),
  width_method = c("original", "middle", "median", "user-defined"),
  height_method = c("original", "middle", "median", "user-defined"),
  length_zero = NA,
  width_zero = NA,
  height_zero = NA,
  ...
)
}
\arguments{
\item{obj_name}{The target viewr object}

\item{axes}{Names of axes to be centered}

\item{length_method}{Method for length}

\item{width_method}{Method for width}

\item{height_method}{Method for height}

\item{length_zero}{User-defined value}

\item{width_zero}{User-defined value}

\item{height_zero}{User-defined value}

\item{...}{Additional arguments}
}
\value{

}
\description{
"Center" the tunnel data, i.e. translation but no rotation
}
