% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rename_viewr_characters}
\alias{rename_viewr_characters}
\title{Quick utility function to use str_replace with mutate(across()) to batch-
rename subjects via pattern detection.}
\usage{
rename_viewr_characters(
  obj_name,
  target_column = "subject",
  pattern,
  replacement = ""
)
}
\arguments{
\item{obj_name}{The target viewr object}

\item{target_column}{The target column; defaults to "subject"}

\item{pattern}{The (regex) pattern to be replaced}

\item{replacement}{The replacement text. Must be a character}
}
\value{

}
\description{
Quick utility function to use str_replace with mutate(across()) to batch-
rename subjects via pattern detection.
}
\seealso{
Other data cleaning functions: 
\code{\link{gather_tunnel_data}()},
\code{\link{get_full_trajectories}()},
\code{\link{quick_separate_trajectories}()},
\code{\link{redefine_tunnel_center}()},
\code{\link{relabel_viewr_axes}()},
\code{\link{rotate_tunnel}()},
\code{\link{select_x_percent}()},
\code{\link{separate_trajectories}()},
\code{\link{standardize_tunnel}()},
\code{\link{trim_tunnel_outliers}()},
\code{\link{visualize_frame_gap_choice}()}
}
\concept{data cleaning functions}
