% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rotate_tunnel}
\alias{rotate_tunnel}
\title{Rotate a tunnel so that perches are approximately aligned}
\usage{
rotate_tunnel(
  obj_name,
  all_heights_min = 0.11,
  all_heights_max = 0.3,
  perch1_len_min = -0.06,
  perch1_len_max = 0.06,
  perch2_len_min = 2.48,
  perch2_len_max = 2.6,
  perch1_wid_min = 0.09,
  perch1_wid_max = 0.31,
  perch2_wid_min = 0.13,
  perch2_wid_max = 0.35,
  ...
)
}
\arguments{
\item{obj_name}{The target viewr object}

\item{all_heights_min}{Minimum perch height}

\item{all_heights_max}{Maximum perch height}

\item{perch1_len_min}{Minimum length value of perch 1}

\item{perch1_len_max}{Maximum length value of perch 1}

\item{perch2_len_min}{Minimum length value of perch 2}

\item{perch2_len_max}{Maximum length value of perch 2}

\item{perch1_wid_min}{Minimum width value of perch 1}

\item{perch1_wid_max}{Maximum width value of perch 1}

\item{perch2_wid_min}{Minimum width value of perch 2}

\item{perch2_wid_max}{Maximum witdh value of perch 2}

\item{...}{Additional arguments that may be passed}
}
\value{

}
\description{
Rotate a tunnel so that perches are approximately aligned
}
