% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{select_x_percent}
\alias{select_x_percent}
\title{Select data in the middle X percent of the length of the tunnel}
\usage{
select_x_percent(obj_name, desired_percent = 33, ...)
}
\arguments{
\item{obj_name}{Target viewr object}

\item{desired_percent}{Measured from the center outwards}

\item{...}{Additional arguments}
}
\value{

}
\description{
Select data in the middle X percent of the length of the tunnel
}
