% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{separate_trajectories}
\alias{separate_trajectories}
\title{Separate rows of data into separately labeled trajectories.}
\usage{
separate_trajectories(
  obj_name,
  max_frame_gap = 1,
  frame_rate_proportion = 0.1,
  frame_gap_messaging = FALSE,
  frame_gap_plotting = FALSE,
  ...
)
}
\arguments{
\item{obj_name}{The input viewr object; a tibble or data.frame with attribute
\code{pathviewR_steps} that includes \code{"viewr"}}

\item{max_frame_gap}{If known, a numeric value to use}

\item{frame_rate_proportion}{If autodetect, proportion of frame rate as
reference}

\item{frame_gap_messaging}{Should frame gaps be reported in the console?}

\item{frame_gap_plotting}{Should frame gap diagnostic plots be shown?}

\item{...}{Additional arguments}
}
\value{

}
\description{
Separate rows of data into separately labeled trajectories.
}
\seealso{
Other data cleaning functions: 
\code{\link{gather_tunnel_data}()},
\code{\link{get_full_trajectories}()},
\code{\link{quick_separate_trajectories}()},
\code{\link{redefine_tunnel_center}()},
\code{\link{relabel_viewr_axes}()},
\code{\link{rename_viewr_characters}()},
\code{\link{rotate_tunnel}()},
\code{\link{select_x_percent}()},
\code{\link{standardize_tunnel}()},
\code{\link{trim_tunnel_outliers}()},
\code{\link{visualize_frame_gap_choice}()}

Other functions that define or clean trajectories: 
\code{\link{get_full_trajectories}()},
\code{\link{quick_separate_trajectories}()},
\code{\link{visualize_frame_gap_choice}()}
}
\concept{data cleaning functions}
\concept{functions that define or clean trajectories}
