% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{standardize_tunnel}
\alias{standardize_tunnel}
\title{Rotate and center a tunnel based on specified landmarks.}
\usage{
standardize_tunnel(
  obj_name,
  landmark_one = "perch1",
  landmark_two = "perch2",
  ...
)
}
\arguments{
\item{obj_name}{The target viewr object}

\item{landmark_one}{Subject name of the first landmark}

\item{landmark_two}{Subject name of the second landmark}

\item{...}{Additional arguments}
}
\value{

}
\description{
Rotate and center a tunnel based on specified landmarks.
}
